apps/backend-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
psycopg[binary]==3.2.4
httpx==0.27.2
